refactor: move `icons` package into `misc`
