+++ src/main/java/module-info.java
@@ -50,9 +50,9 @@ module site.overwrite.auditranscribe {
     exports site.overwrite.auditranscribe.io.exceptions;
     exports site.overwrite.auditranscribe.main_views;
     exports site.overwrite.auditranscribe.main_views.helpers;
-    exports site.overwrite.auditranscribe.main_views.icon;
     exports site.overwrite.auditranscribe.main_views.scene_switching;
     exports site.overwrite.auditranscribe.misc;
+    exports site.overwrite.auditranscribe.misc.icon;
     exports site.overwrite.auditranscribe.music;
     exports site.overwrite.auditranscribe.music.exceptions;
     exports site.overwrite.auditranscribe.music.notes;
@@ -68,9 +68,9 @@ module site.overwrite.auditranscribe {
     opens site.overwrite.auditranscribe to javafx.fxml;
     opens site.overwrite.auditranscribe.main_views to javafx.fxml;
     opens site.overwrite.auditranscribe.main_views.helpers to javafx.fxml;
-    opens site.overwrite.auditranscribe.main_views.icon to javafx.fxml;
     opens site.overwrite.auditranscribe.main_views.scene_switching to javafx.fxml;
     opens site.overwrite.auditranscribe.misc to javafx.fxml;
+    opens site.overwrite.auditranscribe.misc.icon to javafx.fxml;
     opens site.overwrite.auditranscribe.setup_wizard to javafx.fxml;
     opens site.overwrite.auditranscribe.setup_wizard.view_controllers to javafx.fxml;
 }
